Readme updates for QL-800
This seems to work fine on the QL-800, thanks! I've added a tick against
it.

I've also slightly updated the help for `brother_ql_create` to make it
clear that you have to enable the --red option to print on tape which
supports red, otherwise the printer reports an error.

(This happens with the P-Touch software too and it's really unclear what
you're doing wrong.)